+++ packages/globalpayments-js/src/internal/built-in-validations/constants.ts
@@ -1,0 +1,1 @@
+export const BUILT_IN_VALIDATIONS_VALIDATION_ROUND_COUNTER_KEY: string = "validationRoundCounter";